// christmas-task/src/scripts/types/enums.ts
export enum Shapes {
    ball = 'шар',
    snowflake = 'снежинка',
    bell = 'колокольчик',
    cone = 'шишка',
    toy = 'фигурка',
}
export enum Colors {
    white = 'белый',
    yellow = 'желтый',
    red = 'красный',
    blue = 'синий',
    green = 'зелёный',
}
export enum Sizes {
    lSize = 'большой',
    mSize = 'средний',
    sSize = 'малый',
}

export enum TypeFilter {
    shapes = 'shapesFilter',
    colors = 'colorsFilter',
    sizes = 'sizesFilter',
    favorite = 'favoriteFilter',
    count = 'countFilter',
    year = 'yearFilter',
    search = 'searchFilter',
}

export enum TypeSort {
    yearMinMax = 'yearMinMaxSort',
    yearMaxMin = 'yearMaxMinSort',
    nameAZ = 'nameAZSort',
    nameZA = 'nameZASort',
}

export enum TypePage {
    main = 'main',
    filter = 'filter',
    game = 'game',
}

export enum TypeLight {
    rainbow = 'rainbow',
    yellow = 'yellow',
    red = 'red',
    blue = 'blue',
    green = 'green',
    white = 'white',
    none = 'none',
}